GOVSI-683: Move state machine into own package
- Move the state machine from the `helpers` package into it's own `state` package as we are going to create additional classes associated with the state machine.
