+++ shared/src/main/java/uk/gov/di/authentication/shared/state/StateMachine.java
@@ -1,4 +1,4 @@
-package uk.gov.di.authentication.shared.helpers;
+package uk.gov.di.authentication.shared.state;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
